Move Verification type to own package
- Let the type live in its own package
- Add some parsing functions on the type.
